requirements.txt
cffi==1.11.5
ddt==1.1.3
flake8==3.6.0
gevent==1.3.7
greenlet==0.4.15
mccabe==0.6.1
Pillow==5.3.0
PyAutoIt==0.4
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyYAML==3.13
selenium==3.14.1
setuptools==40.6.3
urllib3==1.24
wheel==0.32.3
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
